# public/data/swagger.yaml
openapi: 3.0.0
info:
  title: 'He4rtBot Discord API'
  contact:
    email: hey@danielheart.dev
  version: 0.0.0
paths:
  '/auth/oauth/{provider}':
    post:
      tags:
        - auth
      summary: 'Autenticação de usuário'
      operationId: AuthLogin
      parameters:
        -
          name: code
          in: query
          description: 'Token de autorização da Twitch'
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ...
  /bot/gambling/money:
    put:
      tags:
        - gamification-bot
      summary: 'Atualiza a economia do usuário'
      operationId: PutGamblingMoney
      parameters:
        -
          name: Authorization
          in: header
          description: 'Authorization: he4rt-{key}'
          required: false
          schema:
            type: string
        -
          name: discord_id
          in: query
          description: 'ID do usuário do Discord'
          required: true
          schema:
            type: string
        -
          name: operation
          in: query
          description: 'Add || Reduce'
          required: true
          schema:
            type: string
        -
          name: value
          in: query
          description: 'Valor a ser mutado'
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ...
      security:
        -
          api_key: []
  /bot/gamification/levelup:
    post:
      tags:
        - gamification-bot
      summary: 'Gamification Level Up'
      operationId: PostLevelUp
      parameters:
        -
          name: discord_id
          in: query
          description: 'ID do Discord do usuário'
          required: true
          schema:
            type: string
        -
          name: message
          in: query
          description: 'Mensagem do usuário'
          required: true
          schema:
            type: string
        -
          name: donator
          in: query
          description: 'Se o usuário é um apoiador ou não (apoia-se)'
          required: false
          schema:
            type: string
      responses:
        '200':
          description: ...
      security:
        -
          api_key: []
  /ranking/general:
    get:
      tags:
        - ranking
      summary: 'Retorna o ranking de Level/exp global'
      operationId: getRankingGeneral
      parameters:
        -
          name: type
          in: path
          description: 'Nem existe ainda mas vo deixa documentado'
          required: false
          schema:
            type: string
      responses:
        '200':
          description: ...
  /ranking/messages:
    get:
      tags:
        - ranking
      summary: 'Retorna o ranking de Mensagens'
      operationId: getRankingMessages
      parameters:
        -
          name: type
          in: path
          description: 'week = ultima semana | month = ultimo mes'
          required: false
          schema:
            type: string
      responses:
        '200':
          description: ...
  /users:
    get:
      tags:
        - users
      summary: 'Lista todos os usuários'
      operationId: GetUsers
      parameters:
        -
          name: Authorization
          in: header
          description: 'Authorization: he4rt-{key}'
          required: false
          schema:
            type: string
      responses:
        '200':
          description: ...
      security:
        -
          api_key: []
    post:
      tags:
        - users
      summary: 'Cria um novo usuário'
      operationId: PostUser
      parameters:
        -
          name: Authorization
          in: header
          description: 'Authorization: he4rt-{key}'
          required: false
          schema:
            type: string
        -
          name: discord_id
          in: query
          description: 'ID do usuário do Discord'
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ...
      security:
        -
          api_key: []
  '/users/{discordId}':
    get:
      tags:
        - users
      summary: 'Mostra as informações de um usuário'
      operationId: GetUser
      parameters:
        -
          name: Authorization
          in: header
          description: 'Authorization: he4rt-{key}'
          required: false
          schema:
            type: string
        -
          name: discordId
          in: path
          description: 'ID do usuário do Discord'
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ...
      security:
        -
          api_key: []
    put:
      tags:
        - users
      summary: 'Altera um usuário'
      operationId: PutUser
      parameters:
        -
          name: Authorization
          in: header
          description: 'Authorization: he4rt-{key}'
          required: false
          schema:
            type: string
        -
          name: discordId
          in: path
          description: 'ID do usuário do Discord'
          required: true
          schema:
            type: string
        -
          name: name
          in: query
          description: 'Nome da pessoa'
          required: false
          schema:
            type: string
        -
          name: nickname
          in: query
          description: 'Apelido da pessoa'
          required: false
          schema:
            type: string
        -
          name: git
          in: query
          description: 'Link do git'
          required: false
          schema:
            type: string
        -
          name: about
          in: query
          description: 'Informações pessoais do usuário'
          required: false
          schema:
            type: string
      responses:
        '200':
          description: ...
      security:
        -
          api_key: []
    delete:
      tags:
        - users
      summary: 'Apaga um usuário'
      operationId: DeleteUser
      parameters:
        -
          name: Authorization
          in: header
          description: 'Authorization: he4rt-{key}'
          required: false
          schema:
            type: string
        -
          name: discordId
          in: path
          description: 'ID do usuário do Discord'
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ...
  /users/daily:
    post:
      tags:
        - users
      summary: 'Gerador de hCoins diário'
      operationId: postUserDailyCoins
      parameters:
        -
          name: Authorization
          in: header
          description: 'Authorization: he4rt-{key}'
          required: false
          schema:
            type: string
        -
          name: discord_id
          in: query
          description: 'ID do usuário do Discord'
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ...
  /:
    get:
      description: 'Home page'
      responses:
        default:
          description: 'Welcome page'
components:
  securitySchemes:
    api_key:
      type: apiKey
      name: Authorization
      in: header